Separate and move all math commands to a slash cog
Even though the only maths command in the entire bot is `/squareroot`.
